feat(app): add error boundary on contract verification page
